REF: move platform -> _platform to avoid clash w/ stdlib.
