+++ okonomiyaki/platforms/__init__.py
@@ -1,7 +1,7 @@
 # flake8: noqa
 from .abi import PlatformABI, default_abi
 from .epd_platform import X86, X86_64, EPDPlatform, applies
-from .platform import Platform, OSKind, FamilyKind, NameKind
+from ._platform import Platform, OSKind, FamilyKind, NameKind
 from .python_implementation import PythonABI, PythonImplementation
 
 __all__ = [
